refactor: move logs rendering and fetching into a separate component
